refactor(layout): extract line text joining
Move layout-line span joining into a dedicated helper so line grouping can stay focused on row segmentation. This keeps the CJK, RTL, and semantic-spacing join behavior intact while separating it from gutter detection.
